Delete other Spinners
