+++ package.json
@@ -46,6 +46,7 @@
     "husky": "^8.0.0",
     "postcss": "^8.4.38",
     "prettier": "^3.3.3",
+    "sass": "^1.77.8",
     "tailwindcss": "^3.4.3",
     "vite": "^5.2.10",
     "vite-plugin-eslint": "^1.8.1",
