+++ src/components/Spinner/spinner.module.scss
