--- src/components/Spinner3/Spinner3.tsx
@@ -1,26 +1,0 @@
-import styles from './spinner3.module.scss';
-function Spinner3() {
-	return (
-		<div className={styles.wrapper}>
-			<div className={styles.container}>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-				<div className={styles.circle}></div>
-			</div>
-		</div>
-	);
-}
-export default Spinner3;
